better ros helper. new structure
